Improve logging on sent and received SMS
Add optional logging for sent and received SMS by providing a "--verbose"
argument. Defaults to false to not change behavior.

Also add a "--dry" argument for testing purposes.

Using an "ENTRYPOINT" in the dockerfile instead of a "CMD" allows easy
passing of these arguments via docker-compose.yml or cli.

+++ run.py
@@ -1,11 +1,15 @@
 import os
 
-from flask import Flask, request
+from flask import Flask, request, current_app
 from flask_httpauth import HTTPBasicAuth
 from flask_restful import reqparse, Api, Resource, abort
 
 from support import load_user_data, init_state_machine, retrieveAllSms, deleteSms, encodeSms
 from gammu import GSMNetworks
+
+import argparse
+import logging
+from pprint import pformat
 
 pin = os.getenv('PIN', None)
 ssl = os.getenv('SSL', False)
@@ -16,7 +20,6 @@ machine = init_state_machine(pin)
 app = Flask(__name__)
 api = Api(app)
 auth = HTTPBasicAuth()
-
 
 @auth.verify_password
 def verify(username, password):
@@ -61,7 +64,17 @@ class Sms(Resource):
                 message["SMSC"] = {'Number': args.get("smsc")} if args.get("smsc") else {'Location': 1}
                 message["Number"] = number
                 messages.append(message)
-        result = [machine.SendSMS(message) for message in messages]
+
+        app.logger.debug('Sending message(s): %s', pformat(messages))
+
+        if current_app.config["DRY_RUN"]:
+            app.logger.info("Dry run, will not actually send message.")
+            result = "[OK]"
+        else:
+            result = [machine.SendSMS(message) for message in messages]
+
+        app.logger.debug('Done -- %s', str(result))
+
         return {"status": 200, "message": str(result)}, 200
 
 
@@ -102,11 +115,17 @@ class GetSms(Resource):
         sms = {"Date": "", "Number": "", "State": "", "Text": ""}
         if len(allSms) > 0:
             sms = allSms[0]
-            deleteSms(machine, sms)
+
+            app.logger.debug('Received message: %s', pformat(sms))
+
+            if current_app.config["DRY_RUN"]:
+                app.logger.info("Dry run, will not delete message")
+            else:
+                deleteSms(machine, sms)
+
             sms.pop("Locations")
 
         return sms
-
 
 class SmsById(Resource):
     def __init__(self, sm):
@@ -117,19 +136,44 @@ class SmsById(Resource):
         allSms = retrieveAllSms(machine)
         self.abort_if_id_doesnt_exist(id, allSms)
         sms = allSms[id]
+        app.logger.debug('Received message %s: %s', id, pformat(sms))
         sms.pop("Locations")
         return sms
 
     def delete(self, id):
         allSms = retrieveAllSms(machine)
         self.abort_if_id_doesnt_exist(id, allSms)
-        deleteSms(machine, allSms[id])
+
+        if current_app.config["DRY_RUN"]:
+            app.logger.info("Dry run, will not actually delete message %s", id)
+        else:
+            deleteSms(machine, allSms[id])
+
         return '', 204
 
     def abort_if_id_doesnt_exist(self, id, allSms):
         if id < 0 or id >= len(allSms):
             abort(404, message = "Sms with id '{}' not found".format(id))
 
+def parse_args():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--dry", action="store_true", help="Do not actually send sms")
+    parser.add_argument("--verbose", "-v", action="store_true", help="Log sent and received sms")
+    parser.add_argument("--silent", "-s", action="store_true", help="Suppress logs")
+    return parser.parse_args()
+
+def configure_logging(args):
+    if args.silent:
+        level = logging.ERROR
+    elif args.verbose:
+        level = logging.DEBUG
+    else:
+        level = logging.INFO
+
+    logging.basicConfig(
+        level=level,
+        format="%(asctime)s [%(levelname)s] %(message)s"
+    )
 
 api.add_resource(Sms, '/sms', resource_class_args=[machine])
 api.add_resource(SmsById, '/sms/<int:id>', resource_class_args=[machine])
@@ -139,6 +183,11 @@ api.add_resource(GetSms, '/getsms', resource_class_args=[machine])
 api.add_resource(Reset, '/reset', resource_class_args=[machine])
 
 if __name__ == '__main__':
+
+    args = parse_args()
+    configure_logging(args)
+    app.config["DRY_RUN"] = args.dry
+
     if ssl:
         app.run(port=port, host=host, ssl_context=('/ssl/cert.pem', '/ssl/key.pem'))
     else:
